+++ modules/m_lusers.c
@@ -123,9 +123,10 @@ m_users(struct Client *client_p, struct Client *source_p, int parc, const char *
 	{
 		sendto_one_numeric(source_p, RPL_LOCALUSERS,
 				   form_str(RPL_LOCALUSERS),
-				   rb_dlink_list_length(&lclient_list),
+				   (int)rb_dlink_list_length(&lclient_list),
 				   Count.max_loc,
-				   rb_dlink_list_length(&lclient_list), Count.max_loc);
+				   (int)rb_dlink_list_length(&lclient_list),
+				   Count.max_loc);
 
 		sendto_one_numeric(source_p, RPL_GLOBALUSERS,
 				   form_str(RPL_GLOBALUSERS),
