+++ modules/m_topic.c
@@ -133,7 +133,8 @@ m_topic(struct Client *client_p, struct Client *source_p, int parc, const char *
 
 			sendto_one(source_p, form_str(RPL_TOPICWHOTIME),
 				   me.name, source_p->name, chptr->chname,
-				   chptr->topic->topic_info, chptr->topic->topic_time);
+				   chptr->topic->topic_info,
+				   (unsigned long)chptr->topic->topic_time);
 		}
 	}
